subscriber/requirements.txt
google-cloud-pubsub==2.23.0
google-cloud-spanner==3.47.0

